reorganize universal system with instances folder
