+++ ENIGMAsystem/SHELL/Collision_Systems/BBox/BBOXimpl.cpp
@@ -21,8 +21,8 @@
 ////////////////////////////////////
 
 #include "Universal_System/Object_Tiers/collisions_object.h"
-#include "Universal_System/instance_system.h" //iter
-#include "Universal_System/instance.h"
+#include "Universal_System/Instances/instance_system.h" //iter
+#include "Universal_System/Instances/instance.h"
 
 #include "BBOXutil.h"
 #include "BBOXimpl.h"
