+++ ENIGMAsystem/SHELL/Collision_Systems/Precise/PRECfuncs.cpp
@@ -24,10 +24,10 @@
 #include <limits>
 
 #include "Universal_System/Object_Tiers/collisions_object.h"
-#include "Universal_System/instance_system.h" //iter
+#include "Universal_System/Instances/instance_system.h" //iter
 #include "Universal_System/roomsystem.h"
 #include "Collision_Systems/collision_mandatory.h" //iter
-#include "Universal_System/instance.h"
+#include "Universal_System/Instances/instance.h"
 #include "Universal_System/math_consts.h"
 
 #include "../General/CSfuncs.h"
