+++ ENIGMAsystem/SHELL/Collision_Systems/Precise/PRECimpl.cpp
@@ -22,8 +22,8 @@
 ////////////////////////////////////
 
 #include "Universal_System/Object_Tiers/collisions_object.h"
-#include "Universal_System/instance_system.h" //iter
-#include "Universal_System/instance.h"
+#include "Universal_System/Instances/instance_system.h" //iter
+#include "Universal_System/Instances/instance.h"
 #include "Universal_System/math_consts.h"
 
 #include "PRECimpl.h"
