+++ ENIGMAsystem/SHELL/ENIGMAengine.cbp
@@ -951,8 +951,8 @@
 		<Unit filename="Universal_System/background_internal.h" />
 		<Unit filename="Universal_System/bufferstruct.cpp" />
 		<Unit filename="Universal_System/bufferstruct.h" />
-		<Unit filename="Universal_System/callbacks_events.cpp" />
-		<Unit filename="Universal_System/callbacks_events.h" />
+		<Unit filename="Universal_System/Instances/callbacks_events.cpp" />
+		<Unit filename="Universal_System/Instances/callbacks_events.h" />
 		<Unit filename="Universal_System/Object_Tiers/collisions_object.cpp" />
 		<Unit filename="Universal_System/Object_Tiers/collisions_object.h" />
 		<Unit filename="Universal_System/darray.cpp" />
@@ -978,8 +978,8 @@
 		<Unit filename="Universal_System/highscore_functions.h" />
 		<Unit filename="Universal_System/image_formats.cpp" />
 		<Unit filename="Universal_System/image_formats.h" />
-		<Unit filename="Universal_System/instance.cpp" />
-		<Unit filename="Universal_System/instance.h" />
+		<Unit filename="Universal_System/Instances/instance.cpp" />
+		<Unit filename="Universal_System/Instances/instance.h" />
 		<Unit filename="Universal_System/instance/GLOBAL_GAME_SETTINGS.h" />
 		<Unit filename="Universal_System/instance/IDE_edit_instancode.h" />
 		<Unit filename="Universal_System/instance/INSTANCE_COMMON_VARIABLES.h" />
@@ -987,13 +987,13 @@
 		<Unit filename="Universal_System/instance/cleanup.h" />
 		<Unit filename="Universal_System/instance/events.h" />
 		<Unit filename="Universal_System/instance/instances.h" />
-		<Unit filename="Universal_System/instance_create.h" />
-		<Unit filename="Universal_System/instance_iterator.h" />
-		<Unit filename="Universal_System/instance_planar.cpp" />
-		<Unit filename="Universal_System/instance_system.cpp" />
-		<Unit filename="Universal_System/instance_system.h" />
-		<Unit filename="Universal_System/instance_system_base.h" />
-		<Unit filename="Universal_System/instance_system_frontend.h" />
+		<Unit filename="Universal_System/Instances/instance_create.h" />
+		<Unit filename="Universal_System/Instances/instance_iterator.h" />
+		<Unit filename="Universal_System/Instances/instance_planar.cpp" />
+		<Unit filename="Universal_System/Instances/instance_system.cpp" />
+		<Unit filename="Universal_System/Instances/instance_system.h" />
+		<Unit filename="Universal_System/Instances/instance_system_base.h" />
+		<Unit filename="Universal_System/Instances/instance_system_frontend.h" />
 		<Unit filename="Universal_System/key_game_globals.h" />
 		<Unit filename="Universal_System/lives.cpp" />
 		<Unit filename="Universal_System/lives.h" />
@@ -1044,7 +1044,7 @@
 		<Unit filename="Universal_System/var4.h" />
 		<Unit filename="Universal_System/var4_lua.cpp" />
 		<Unit filename="Universal_System/var_te.h" />
-		<Unit filename="Universal_System/with.h" />
+		<Unit filename="Universal_System/Instances/with.h" />
 		<Unit filename="Universal_System/zlib.cpp" />
 		<Unit filename="Universal_System/zlib.h" />
 		<Unit filename="Widget_Systems/None/Info/widget_info.h">
