+++ ENIGMAsystem/SHELL/Graphics_Systems/General/actions.h
@@ -30,7 +30,7 @@
 
 #include "Universal_System/scalar.h"
 #include "Universal_System/GAME_GLOBALS.h" // TODO: Do away with this sloppy infestation permanently!
-#include "Universal_System/instance_system_base.h"
+#include "Universal_System/Instances/instance_system_base.h"
 #include "Universal_System/lives.h"
 
 namespace enigma_user
