+++ ENIGMAsystem/SHELL/Graphics_Systems/None/fillin.cpp
@@ -40,7 +40,7 @@
 #include "Universal_System/nlpo2.h"
 #include "Universal_System/fileio.h"
 #include "Universal_System/estring.h"
-#include "Universal_System/instance_system.h"
+#include "Universal_System/Instances/instance_system.h"
 #include "Universal_System/Object_Tiers/graphics_object.h"
 #include "Universal_System/estring.h"
 #include "Platforms/platforms_mandatory.h"
