+++ ENIGMAsystem/SHELL/SHELLmain.cpp
@@ -41,7 +41,7 @@
 #include "Universal_System/fonts.h"
 #include "Universal_System/texture_atlas.h"
 
-#include "Universal_System/callbacks_events.h"
+#include "Universal_System/Instances/callbacks_events.h"
 
 #include "GameSettings.h"
 #include "Preprocessor_Environment_Editable/LIBINCLUDE.h"
@@ -71,12 +71,12 @@ extern int amain();
 #include "Universal_System/image_formats.h"
 
 #include "Universal_System/Object_Tiers/object.h"
-#include "Universal_System/instance.h"
+#include "Universal_System/Instances/instance.h"
 #include "Universal_System/roomsystem.h"
 
 #include "Universal_System/globalupdate.h"
 
-#include "Universal_System/instance_system_frontend.h"
+#include "Universal_System/Instances/instance_system_frontend.h"
 
 #include "Universal_System/resource_data.h"
 #include "Universal_System/highscore_functions.h"
@@ -96,7 +96,7 @@ using namespace enigma_user;
   #ifndef JUST_DEFINE_IT_RUN
   #include "Universal_System/syntax_quirks.h"
 
-  #include "Universal_System/with.h"
+  #include "Universal_System/Instances/with.h"
   #include "Preprocessor_Environment_Editable/IDE_EDIT_evparent.h"
   #include "Preprocessor_Environment_Editable/IDE_EDIT_events.h"
   #include "Preprocessor_Environment_Editable/IDE_EDIT_objectdeclarations.h"
@@ -113,7 +113,7 @@ using namespace enigma_user;
     #include "Modes/buildmode.h"
   #endif
 
-  #include "Universal_System/instance_create.h"
+  #include "Universal_System/Instances/instance_create.h"
 #endif
 
 namespace enigma
