+++ ENIGMAsystem/SHELL/Universal_System/Extensions/MotionPlanning/actions.h
@@ -26,7 +26,7 @@
  **                                                                              **
  \********************************************************************************/
 
-#include "Universal_System/instance_system_base.h"
+#include "Universal_System/Instances/instance_system_base.h"
 
 namespace enigma_user
 {
