+++ ENIGMAsystem/SHELL/Universal_System/Extensions/MotionPlanning/mp_movement.cpp
@@ -30,7 +30,7 @@
 
 #include "Universal_System/scalar.h"
 #include "Universal_System/Object_Tiers/collisions_object.h"
-#include "Universal_System/instance_system.h"
+#include "Universal_System/Instances/instance_system.h"
 #include "Universal_System/math_consts.h"
 
 #include "../Paths/pathstruct.h"
