+++ ENIGMAsystem/SHELL/Universal_System/Extensions/Paths/path_functions.cpp
@@ -48,7 +48,7 @@ namespace enigma {
 }
 
 #include "Universal_System/Object_Tiers/collisions_object.h"
-#include "Universal_System/instance_system.h"
+#include "Universal_System/Instances/instance_system.h"
 #include "implement.h"
 
 namespace enigma {
