+++ ENIGMAsystem/SHELL/Universal_System/Instances/instance_system.h
@@ -23,9 +23,9 @@
 #define ENIGMA_INSTANCE_SYSTEM_H
 
 #include "instance_iterator.h"
-#include "Object_Tiers/object.h"
-#include "reflexive_types.h"
-#include "var4.h"
+#include "Universal_System/Object_Tiers/object.h"
+#include "Universal_System/reflexive_types.h"
+#include "Universal_System/var4.h"
 
 #include <map>
 #include <set>
