+++ ENIGMAsystem/SHELL/Universal_System/Instances/instance_system_base.h
@@ -19,7 +19,7 @@
 #ifndef INSTANCE_SYSTEM_BASE_h
 #define INSTANCE_SYSTEM_BASE_h
 
-#include "Object_Tiers/object.h"
+#include "Universal_System/Object_Tiers/object.h"
 #include <string>
 
 namespace enigma
