+++ ENIGMAsystem/SHELL/Universal_System/Makefile
@@ -1,3 +1,3 @@
-SOURCES += $(wildcard Universal_System/*.cpp) $(wildcard Universal_System/Object_Tiers/*.cpp)
+SOURCES += $(wildcard Universal_System/*.cpp) $(wildcard Universal_System/Object_Tiers/*.cpp) $(wildcard Universal_System/Instances/*.cpp)
 override LDLIBS += -lz
 CLEANDEPS += $(OBJDIR)/Universal_System/loading.o
