+++ ENIGMAsystem/SHELL/Universal_System/Object_Tiers/object.cpp
@@ -23,7 +23,7 @@
 #include "libEGMstd.h"
 
 #ifdef DEBUG_MODE
-  #include "Universal_System/instance_system.h"
+  #include "Universal_System/Instances/instance_system.h"
   #include "Universal_System/resource_data.h" // TODO: We don't guarantee these functions exist. But they're useful for debugging. We need a debug namespace that offers this, too.
   #include <Widget_Systems/widgets_mandatory.h> // show_error
 #endif
