+++ ENIGMAsystem/SHELL/Universal_System/depth_draw.h
@@ -35,7 +35,7 @@
 #ifndef ENIGMA_DEPTH_DRAW_H
 #define ENIGMA_DEPTH_DRAW_H
 
-#include "instance_system.h"
+#include "Instances/instance_system.h"
 #include "roomsystem.h"
 
 #include <map>
