+++ ENIGMAsystem/SHELL/Universal_System/move_functions.cpp
@@ -20,7 +20,7 @@
 #include <floatcomp.h>
 
 #include "Object_Tiers/collisions_object.h"
-#include "instance_system.h"
+#include "Instances/instance_system.h"
 #include "roomsystem.h"
 #include "move_functions.h"
 #include "math_consts.h"
